\name{framingham}
\docType{data}
\alias{framingham}
\usage{data(framingham)}
\title{Framingham Data}
\description{
The data is from the Framingham Study on coronary heart disease described by Carroll et al. (2006). The data consist of measurements of systolic blood pressure (SBP) obtained at two different examinations in 1,615 males on an 8-year follow-up from the first examination. At each examination, the SBP was measured twice and for each individual.
}
\format{A data frame with 1615 observations on 4 variables.
  \tabular{lll}{
    \code{SBP11}  \tab numeric  \tab First systolic blood pressure at exam 1 \cr
    \code{SBP12}   \tab numeric  \tab Second systolic blood pressure at exam 1 \cr
    \code{SBP21}  \tab numeric  \tab First systolic blood pressure at exam 2 \cr
    \code{SBP22}   \tab numeric  \tab Second systolic blood pressure at exam 2 \cr
  }
}
\seealso{\code{\link{DeconPdf}}}
\references{
Carroll, R.~J., Ruppert, D., Stefanski, L.~A., and Crainiceanu, C. (2006).
\emph{Measurement Error in Nonlinear Models: A Modern Perspective,
  Second Edition}. Chapman Hall, New York.
}
\keyword{datasets}
