% This is a template R document file
% Distributed under GPL 3 or later

\name{plot.Decon}
\alias{plot.Decon}
\title{Plot a Decon Object}
\description{
  To plot a ``Decon'' object generated by DeconPdf(...), DeconCdf(...), DeconCPdf(...), or DeconNpr(...).
}
\details{
This function is to plot the estimated function generated by DeconPdf(...), DeconCdf(...), DeconCPdf(...), or DeconNpr(...)}
\author{
  X.F. Wang \email{wangx6@ccf.org}

  B. Wang  \email{bwang@jaguar1.usouthal.edu}
}
\seealso{
 \code{\link{DeconPdf}, \link{DeconCdf}, \link{DeconCPdf}, \link{DeconNpr}}.
}
\keyword{smooth}

