% This is a template R document file
% Distributed under GPL 3 or later

\name{plot.DeconCdf}
\alias{plot.DeconCdf}
\title{Plot DeconCdf Objects}
\description{
  To plot objects generated by DeconCdf(...)
}
\details{
This function is to plot the estimated CDF by DeconCdf}
\author{
  X.F. Wang \email{wangx6@ccf.org}

  B. Wang  \email{bwang@jaguar1.usouthal.edu}
}
\seealso{
 \code{\link{DeconCdf}}.
}
\keyword{smooth}

