% This is a template R document file
% Distributed under GPL 3 or later

\name{plot.DeconCPdf}
\alias{plot.DeconCPdf}
\title{Plot DeconCPdf Objects}
\description{
  To plot objects generated by DeconCPdf(...)
}
\details{
This function is to plot the estimated conditional PDF by DeconCPdf}
\author{
  X.F. Wang \email{wangx6@ccf.org}

  B. Wang  \email{bwang@jaguar1.usouthal.edu}
}
\seealso{
 \code{\link{DeconCPdf}}.
}
\keyword{smooth}

