% This is a template R document file
% Distributed under GPL 3 or later

\name{plot.DeconNpr}
\alias{plot.DeconNpr}
\title{Plot DeconNpr Objects}
\description{
  To plot objects generated by DeconNpr(...)
}
\details{
This function is to plot the estimated regression function by DeconNpr}
\author{
  X.F. Wang \email{wangx6@ccf.org}

  B. Wang  \email{bwang@jaguar1.usouthal.edu}
}
\seealso{
 \code{\link{DeconNpr}}.
}
\keyword{smooth}

