% This is a template R document file
% Distributed under GPL 3 or later

\name{plot.DeconPdf}
\alias{plot.DeconPdf}
\title{Plot DeconPdf Objects}
\description{
  To plot objects generated by DeconPdf(...)
}
\details{
This function is to plot the estimated pdf by DeconPdf}
\author{
  X.F. Wang \email{wangx6@ccf.org}

  B. Wang  \email{bwang@jaguar1.usouthal.edu}
}
\seealso{
 \code{\link{DeconPdf}}.
}
\keyword{smooth}

