% This is a template R document file
% Distributed under GPL 3 or later

\name{print.DeconCdf}
\alias{print.DeconCdf}
\title{Print DeconCdf Objects}
\description{
  To print objects generated by DeconCdf(...)
}
\details{
This function is to print the DeconCdf object generated  by DeconCdf}
\author{
  X.F. Wang \email{wangx6@ccf.org}

  B. Wang  \email{bwang@jaguar1.usouthal.edu}
}
\seealso{
 \code{\link{DeconCdf}}.
}
\keyword{smooth}

