% This is a template R document file
% Distributed under GPL 3 or later

\name{print.DeconCPdf}
\alias{print.DeconCPdf}
\title{Print DeconCPdf Objects}
\description{
  To print objects generated by DeconCPdf(...)
}
\details{
This function is to print the DeconCPdf object generated  by DeconCPdf}
\author{
  X.F. Wang \email{wangx6@ccf.org}

  B. Wang  \email{bwang@jaguar1.usouthal.edu}
}
\seealso{
 \code{\link{DeconCPdf}}.
}
\keyword{smooth}

