% This is a template R document file
% Distributed under GPL 3 or later

\name{print.DeconPdf}
\alias{print.DeconPdf}
\title{Print DeconPdf Objects}
\description{
  To print objects generated by DeconPdf(...)
}
\details{
This function is to print the DeconPdf object generated  by DeconPdf}
\author{
  X.F. Wang \email{wangx6@ccf.org}

  B. Wang  \email{bwang@jaguar1.usouthal.edu}
}
\seealso{
 \code{\link{DeconPdf}}.
}
\keyword{smooth}

